:recycle:add: new route /sale do post new sale
