+++ back-end/src/controllers/index.js
@@ -1,7 +1,9 @@
 const userController = require('./userController');
 const productController = require('./productController');
+const saleController = require('./saleController');
 
 module.exports = {
   userController,
   productController,
+  saleController,
 };
