+++ back-end/src/routers/index.js
@@ -1,9 +1,11 @@
 const loginRouter = require('./loginRouter');
 const productRouter = require('./productRouter');
 const registerRouter = require('./registerRouter');
+const saleRouter = require('./saleRouter');
 
 module.exports = { 
   loginRouter,
   registerRouter,
   productRouter,
+  saleRouter,
 };
