+++ back-end/src/services/index.js
@@ -1,7 +1,9 @@
 const userService = require('./userService');
 const productService = require('./productService');
+const saleService = require('./saleService');
 
 module.exports = {
   userService,
   productService,
+  saleService,
 };
